+++ src/main/resources/db/migration/V31__create_table_supply_payments.sql
@@ -6,7 +6,7 @@ CREATE TABLE supply_payments (
     payment_to VARCHAR(100) NOT NULL,
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL,
-    version INTEGER DEFAULT 0 NOT NULL,
+    version BIGINT DEFAULT 0 NOT NULL,
 
     CONSTRAINT fk_supply_payments FOREIGN KEY (supply_id)
         REFERENCES supplies(id)
